src/test/java/com/vulniq/client/securityanalyzer/datacollection/docker/DockerInfoCollectorTest.java


/*
 *     VulnIQ Security Scanner, Terzi
 *     Copyright (C) 2021  SBOSoft A.S. and its affiliates
 *     Contact : info@vulniq.com www.vulniq.com
 *
 *     This program is free software: you can redistribute it and/or modify
 *     it under the terms of the GNU General Public License as published by
 *     the Free Software Foundation, either version 3 of the License, or
 *     (at your option) any later version.
 *
 *     This program is distributed in the hope that it will be useful,
 *     but WITHOUT ANY WARRANTY; without even the implied warranty of
 *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *     GNU General Public License for more details.
 *
 *     You should have received a copy of the GNU General Public License
 *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

package com.vulniq.client.securityanalyzer.datacollection.docker;

import com.google.gson.Gson;
import com.google.gson.GsonBuilder;
import com.vulniq.client.securityanalyzer.TerziTestBase;
import com.vulniq.client.securityanalyzer.TerziTestSetup;
import com.vulniq.client.securityanalyzer.TerziTestUtilizingDocker;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

import java.util.Date;
import java.util.Map;

public class DockerInfoCollectorTest extends TerziTestUtilizingDocker
{

    @Test
    public void dockerInspect()
    {
        String s = "terzi-debian10";
        Map<String, Object> d = DockerInfoCollector.dockerInspect(s);
        //System.out.println(d);
        Assert.assertNotNull(d);
    }

    @Test
    public void dockerPS()
    {
        Map<String, DockerContainerInfo> containers = DockerInfoCollector.dockerPS();
        //System.out.println(containers);
        Assert.assertNotNull(containers);
    }


}